feat: 마이페이지 프로젝트 컴포넌트 분리
